// packages/reason-editor/src/types.ts
export interface GeneralOptions<T> {
  /** Enabled divider */
  divider: boolean;
  /** Enabled spacer */
  spacer: boolean;
  /** Button view function */
  button: ButtonView<T>;
  /** Show on Toolbar */
  toolbar?: boolean;
  /** Shortcut keys override */
  shortcutKeys?: string[] | string[][];
}
// ...
export interface ButtonViewReturn {
  /** Component */
  component: unknown;
  /** Component props */
  componentProps: ButtonViewReturnComponentProps;
  /** Component slots */
  componentSlots?: ButtonViewReturnComponentSlots;
}

/**
 * Represents the parameters for the ButtonView function.
 */
export interface ButtonViewParams<T = any> {
  /** Editor object */
  editor: Editor;
  /** Extension object */
  extension: Extension<T>;
  /** Translation function */
  t: (path: string) => string;
}

/**
 * Represents the ButtonView function.
 */
export type ButtonView<T = any> = (
  options: ButtonViewParams<T>
) => ButtonViewReturn | ButtonViewReturn[];